add guess_word.py, update game.py, methods from hidden_words replace to game.py
