+++ bot_viselnica.py
@@ -1,6 +1,5 @@
 from telegram import Update
 from telegram.ext import Updater, CommandHandler, CallbackContext
-import hidden_word as hd
 import game as gm
 
 def hello(update: Update, context: CallbackContext) -> None:
@@ -10,7 +9,7 @@ def hello(update: Update, context: CallbackContext) -> None:
 updater = Updater('5311753371:AAE_hvFU2BT3XtKv5yvuMj45bZULxT8u0kc')
 
 updater.dispatcher.add_handler(CommandHandler('hello', hello))
-updater.dispatcher.add_handler(CommandHandler('topic', hd.hidden_word_topic))
+updater.dispatcher.add_handler(CommandHandler('topic', gm.hidden_word_topic))
 updater.dispatcher.add_handler(CommandHandler('game', gm.game))
 print('server_started')
