+++ game.py
@@ -1,22 +1,59 @@
 from dataclasses import replace
 from telegram import Update
 from telegram.ext import Updater, CommandHandler, CallbackContext
-import hidden_word as hd
+import random
 import re
+from guess_word import word_competion
 
 
-word = hd.word()
+whattopic = random.randint(1, 4)
+
+def hidden_word_topic(update: Update, context: CallbackContext):
+    global censured_word
+    if whattopic == 1:
+        update.message.reply_text('The topic is ANIMALS')
+    elif whattopic == 2:
+        update.message.reply_text('The topic is FRUITS')
+    elif whattopic == 3:
+        update.message.reply_text('The topic is CONTRIES')
+    update.message.reply_text(f'{censured_word}, please enter /game (space) a letter you want to guess')
+
+def word_choise():
+    list1 = ['cat','dog','pig']
+    list2 = ['apple', 'banana', 'orange']
+    list3 = ['spain', 'france', 'germany']
+    if whattopic == 1:
+        i = random.randint(0, 2)
+        word = list1[i]
+        return word
+    elif whattopic == 2:
+        i = random.randint(0, 2)
+        word = list2[i]
+        return word
+    elif whattopic == 3:
+        i = random.randint(0, 2)
+        word = list3[i]
+        return word
+
+word = word_choise()
 censured_word = re.sub('[a-z]','*',word)
 count = 8
 
 def game(update: Update, context: CallbackContext):
     global count
+    global censured_word
+    
     while True:
-        update.message.reply_text(f'This is the word you should guess {censured_word}, please enter /game (space) a letter you want to guess')
         msg = update.message.text
-        reslist = msg.split()
+        reslist = msg.split()        
         if reslist[1] in word:
-            update.message.reply_text(f'Your letter ({reslist[1]}) is inside the word')
+            word_user = word_competion(reslist[1], censured_word, word)
+            censured_word = word_user
+            if censured_word.find('*') != -1:
+                update.message.reply_text(f'{censured_word}, please enter /game (space) a letter you want to guess')
+            else:
+                update.message.reply_text('Congratulations! You are win! \n Do you want to guess again?')
+            break    
         else:
             update.message.reply_text(f'Your letter ({reslist[1]}) is NOT inside your word')
             count = count - 1
@@ -46,6 +83,9 @@ def game(update: Update, context: CallbackContext):
                 update.message.reply_text('You dont have more lifes, you are hanged')
                 break
 
+
+def again(update: Update, context: CallbackContext):
+    update.message.reply_text('Do you want to guess again?')
